refactor: apollo client
