+++ package.json
@@ -75,6 +75,9 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.16",
+    "@n1ru4l/graphql-live-query-patch-json-patch": "^0.6.0",
+    "@n1ru4l/push-pull-async-iterable-iterator": "^3.1.0",
+    "@n1ru4l/socket-io-graphql-client": "^0.11.1",
     "@nestjs/common": "^8.2.3",
     "@nestjs/core": "^8.2.3",
     "@nestjs/graphql": "^9.1.2",
